requirements.txt
langchain
openai
streamlit
faiss-cpu
pypdf
tiktoken
python-dotenv